requirements.txt
gym[mujoco,atari,classic_control,robotics]
demjson==2.2.4
joblib==0.12.0
pandas==0.22.0
matplotlib==3.0.2
joblib==0.12.0
tensorflow_gpu==1.10.1
cloudpickle==0.5.3
opencv_python==3.4.0.12
scipy==1.1.0
numpy==1.14.1
gym==0.10.4
pathos==0.2.2.1
dotmap==1.2.20